move whatsapp actions into namespaced module
