+++ src/api/appType/whatsapp/index.js
@@ -1,3 +1,4 @@
+import qs from 'query-string';
 import request from '@/api/request.js';
 
 const resource = '/api/v1/apptypes';
@@ -7,5 +8,24 @@ export default {
   },
   updateWppContactInfo(appCode, appUuid, payload) {
     return request.$http.patch(`${resource}/${appCode}/apps/${appUuid}/contact/`, payload);
+  },
+  getSharedWabas(appCode, filter) {
+    const queryString = qs.stringify(filter);
+    return request.$http.get(`${resource}/${appCode}/apps/shared-wabas/?${queryString}`);
+  },
+  getConversations(appCode, appUuid, params) {
+    const queryString = qs.stringify(params);
+    return request.$http.get(
+      `${resource}/${appCode}/apps/${appUuid}/conversations/?${queryString}`,
+    );
+  },
+  fetchWppProfile(appCode, appUuid) {
+    return request.$http.get(`${resource}/${appCode}/apps/${appUuid}/profile/`);
+  },
+  updateWppProfile(appCode, appUuid, data) {
+    return request.$http.patch(`${resource}/${appCode}/apps/${appUuid}/profile/`, data);
+  },
+  deleteWppProfilePhoto(appCode, appUuid) {
+    return request.$http.delete(`${resource}/${appCode}/apps/${appUuid}/profile/`);
   },
 };
