+++ src/store/appType/actions.js
@@ -44,24 +44,4 @@ export default {
   async updateAppConfig(store, { code, appUuid, payload }) {
     return await appType.updateAppConfig(code, appUuid, payload);
   },
-
-  async getSharedWabas(store, { code, params }) {
-    return await appType.getSharedWabas(code, params);
-  },
-
-  async getConversations(store, { code, appUuid, params }) {
-    return await appType.getConversations(code, appUuid, params);
-  },
-
-  async fetchWppProfile(store, { code, appUuid }) {
-    return await appType.fetchWppProfile(code, appUuid);
-  },
-
-  async updateWppProfile(store, { code, appUuid, payload }) {
-    return await appType.updateWppProfile(code, appUuid, payload);
-  },
-
-  async deleteWppProfilePhoto(store, { code, appUuid }) {
-    return await appType.deleteWppProfilePhoto(code, appUuid);
-  },
 };
