+++ src/store/appType/channels/whatsapp/actions.js
@@ -15,4 +15,19 @@ export default {
     const { data } = await whatsApp.updateWppContactInfo(code, appUuid, payload);
     commit('SET_WPP_CONTACT_INFO', data);
   },
+  async getSharedWabas(store, { code, params }) {
+    return await whatsApp.getSharedWabas(code, params);
+  },
+  async getConversations(store, { code, appUuid, params }) {
+    return await whatsApp.getConversations(code, appUuid, params);
+  },
+  async fetchWppProfile(store, { code, appUuid }) {
+    return await whatsApp.fetchWppProfile(code, appUuid);
+  },
+  async updateWppProfile(store, { code, appUuid, payload }) {
+    return await whatsApp.updateWppProfile(code, appUuid, payload);
+  },
+  async deleteWppProfilePhoto(store, { code, appUuid }) {
+    return await whatsApp.deleteWppProfilePhoto(code, appUuid);
+  },
 };
